pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "seraph-mcp"
version = "1.0.0"
description = "Comprehensive AI Optimization Platform - Token optimization, model routing, semantic caching, and cost management for LLM APIs"
readme = "README.md"
authors = [
    {name = "Seraph Team", email = "team@seraph-mcp.dev"}
]
license = "MIT"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
requires-python = ">=3.12"

# All dependencies included in main package (monolithic architecture)
dependencies = [
    # Core MCP server and configuration
    "fastmcp>=2.0.0", # MCP stdio server
    "pydantic>=2.0.0", # Typed configuration and validation
    "pydantic-settings>=2.0.0", # Settings management
    "python-dotenv>=1.0.0", # Environment variables
    "httpx>=0.25.0", # HTTP client for Models.dev API
    # Cache and storage
    "redis>=5.0.0", # Redis backend for caching
    "cachetools>=5.3.0", # In-memory caching utilities (LRU, FIFO, TTL)
    # AI Model Providers (all included)
    "openai>=1.0.0", # OpenAI GPT models
    "anthropic>=0.25.0", # Anthropic Claude models
    "google-genai>=0.2.0", # Google Gemini models (newer SDK)
    # Token optimization
    "tiktoken>=0.5.0", # OpenAI token counting
    "llmlingua>=0.2.2",
    "blake3>=1.0.7",
    # Resilience patterns
    "pybreaker>=1.0.0", # Circuit breaker pattern
]

[project.optional-dependencies]
# Semantic cache dependencies (optional - needed for semantic caching feature)
semantic_cache = [
    "chromadb>=0.4.0",               # Vector database for similarity search
]

# Development dependencies
dev = [
    "pytest>=8.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "coverage>=7.0.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "types-redis>=4.6.0",
    "types-cachetools>=5.3.0",
    "bandit[toml]>=1.7.0",
    "safety>=3.0.0",
    "pre-commit>=3.5.0",
    "build>=1.0.0",
    "twine>=4.0.0",
]

# All optional features (for convenience)
all = [
    "chromadb>=0.4.0",               # Semantic cache
]

[project.scripts]
seraph-mcp = "src.server:main"

[project.urls]
Repository = "https://github.com/coderdayton/seraph-mcp"
Documentation = "https://seraph-mcp.readthedocs.io"
Issues = "https://github.com/coderdayton/seraph-mcp/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "UP"]
ignore = ["E501", "B008", "C901"]  # Ignore line length, assert statements, and complexity

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = "src.server"
warn_unreachable = false

[[tool.mypy.overrides]]
module = "src.providers.gemini_provider"
warn_unreachable = false

[[tool.mypy.overrides]]
module = "llmlingua"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "cachetools"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "chromadb"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "chromadb.config"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "src.semantic_cache.*"
follow_imports = "skip"

[[tool.mypy.overrides]]
module = "src.context_optimization.seraph_compression"
follow_imports = "skip"



[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
pythonpath = ["."]
addopts = [
    "-v",
    "--strict-markers",
    "--strict-config",
    "--showlocals",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow tests",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/site-packages/*",
    "*/server.py",  # MCP server entry point - tested via integration, not unit tests
    "*/providers/*_provider.py",  # Provider implementations require external API access
    "*/providers/factory.py",  # Provider factory requires external APIs
    "*/providers/openai_compatible.py",  # OpenAI-compatible requires external APIs
    "*/providers/models_dev.py",  # Models.dev API requires external access
    "*/semantic_cache/*",  # Optional feature requiring ChromaDB
    "*/budget_management/*",  # Complex module with external dependencies
    "*/context_optimization/middleware.py",  # Requires AI providers
    "*/context_optimization/optimizer.py",  # Requires AI providers
    "*/observability/monitoring.py",  # Observability requires external monitoring setup
    "*/cache/backends/redis.py",  # Redis backend requires Redis server
    "*/__init__.py",  # Import-only files
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
fail_under = 70
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.bandit]
exclude_dirs = ["tests", "venv", ".venv", "env"]
skips = ["B101", "B601", "B311"]  # B311: random is safe for retry jitter (non-cryptographic)

[tool.basedpyright]
typeCheckingMode = "strict"
include = ["src"]
exclude = ["tests", "venv", ".venv", "env", "__pycache__", "build", "dist"]
reportInvalidCast = true
reportUnreachable = true
reportMissingTypeStubs = false
reportUnknownVariableType = false
reportUnknownArgumentType = false
reportUnknownMemberType = false
reportConstantRedefinition = false
reportUnnecessaryIsInstance = false
